Added draw.io diagram
Signal for Trondheim side is mostly charted, but it doesn't show the Oslo connections. Using draw.io to open the file it is possible to edit the diagram. It can then be exported image files. It should be pretty easy to add figures and text, but if there is any problem, I can be contacted.